added ability to choose to keep the file, fixed readme and added .gitignore

+++ quickplay.sh
@@ -1,6 +1,15 @@
 echo "enter your search:"
 read msg
+echo "do you want to keep the file? [yes/no]"
+read choice
 youtube-dl "ytsearch: $msg" -o '%(title)s.%(ext)s' --format mp4
 vid=$(youtube-dl "ytsearch: $msg" --get-filename -o '%(title)s.%(ext)s'  --format mp4)
 mpv "$vid"
-rm -rf "$vid"
+
+if [ "$choice" = "yes" ]
+then
+    echo "DONE"
+else
+    rm -rf "$vid"
+    echo "DONE"
+fi
